Rename pageView to pageCollapse to prevent it being blocked
uBlock and other adblockers think 'pageView.js' is an analytics file,
and thus block it. By renaming it, it will no longer be blocked.

The links for expanding tabs in the wiki sidebar have had their hrefs
removed, as it was making the browser jump to '#' which would move it
to the top of the page on mobile.

Fixes #297

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>