Show who Signed a Plugin in the Versions List
Since Projects can have multiple authors, I've added an authorId field
to Versions (and updated the DB). When someone publishes a Version,
they are assigned as the author.

A key appears in the Versions list next to Versions that have a proper
authorId. If they don't have an authorId (like all the ones currently
in the DB), there won't be a key. The key links to author's page.

Closes #300

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>