Trim usernames when modifying Organization Memberships
There was some whitespace being passed into the equals check, causing
it to fail.

Members that have been invited will now display what role they will be,
and that role is editable.

An NPE with Org PGP Keys was fixed by checking to see if it exists
before using it.

Fixes #301

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>