Start on Tags System
isSpongePlugin and isForgeMod have been converted into Tags. Other Tags
will exist later, but for now we only have Sponge and Forge Tags.

Migrating the DB to the new format will be the toughest part. I have
started on the SQL to convert it, but it needs to be reviewed & tested.

Internally, the code is ready for the new Tags. All spots of isSponge/
isForge usage have been changed to use the new Tags.

Signed-off-by: Jadon Fowler <jadonflower@gmail.com>